requirements.txt
vt
pydantic
pyjwt
starlette
fastapi
uvicorn
slowapi
bcrypt
python-jose
jproperties
pymongo
